Add tini to Dockerfile and use it as entrypoint (#606)
This change adds [tini](https://github.com/krallin/tini) as the init
process for the Teslamate Docker container. The reason for this is that
currently the container runs with zombie processes.

This means the current process with PID=1 in the container doesn't reap
children properly and given the output I get while fixing this, child
processes are dying accidentally.

There's also an option to work around this with docker-compose since
compose file version 3.7, which allows passing `--init` for the docker
container. I'm currently using that to work around the zombie processes,
but adding tini fixes this independent of how people run this container.