Display marketing names for Model S and X
All stemming from #2494 - I am going to make this pull request fully knowing it either will be rejected (totally OK) or require some additional work to get it working.

I am not at all close to knowing where the `car.model` or `car.trim_badging` comes from. But I do know to get the right Model S designations, we need to know the `car.car_type`. As such, we need to add this into the logic for lines 64 through 71 (before changes). Also need to add it to line 61 (I believe) to check for this data. *However*, I have no clue what is going to happen to the Model 3 and Y logic that is there if we do not add the `car_type` logic to it. Or we just add repetitive checks, because `car.model` already equals 3 and `car.car_type` also equals 3 (per this comment https://github.com/adriankumpf/teslamate/issues/2494#issuecomment-1049360750)

But to get the right Model S (And X) names, we need not only the `car.model`, but since they started naming models with the 2021.5 refresh, we need the `car.car_type` attribute as well. So I took at stab at it.