Add `charge_rate` to MQTT broker
I'm switching from the Home Assistant Tesla integration to TeslaMate. I had a dashboard where I displayed the `charge_rate`. This was the only value I was missing in the mqtt topics from TeslaMate.

I was wondering if there was a reason why it was not exported?
It is covered by the vehicle state api call:
https://github.com/adriankumpf/teslamate/blob/7a9c6b9ad1b2b03f424f46f652a53b831df73697/lib/tesla_api/vehicle/state.ex#L38

I'd love to see this added.