Fix link to Charge Level dashboard
I noticed this link that should probably go to the charge level dashboard.

Also, semi-related, is there a reason all the uid's are random numbers? These are the uid's that appear in the URL of the dasboard. I changed them on my instance to more readable uid's, is this something that would be approved if I changed all the uid's of the dashboards, and fix all the links? The only breaking change I can foresee is bookmarked dashboards become invalid.