dashboards: re-enable time picker by default in drive-details, charge-details
While the time picker doesn't work quite correctly with internal
dashboards (i.e., not meant to be called independently, which will
cause N/A data errors), it is still useful to have it enabled:

It provides at-a-glance from-to time at the top bar for easy reference.
When you call it independently, it allows you to monitor real-time stats
(e.g., Using "Last 1 hour" with auto refresh).

Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>