docs: improvement to the Debian installation instructions (#4734)
* Slight improvement to the Debian installation instructions

I've managed to do an install on Debian Trixie with most of the required
dependencies from Debian, with the exception of Grafana (which is not)
packaged. As such, I've modified the instructions to reflect this.

It's possible that in the future, TeslaMate will require newer versions,
and these instructions would become outdated. But that's not worse than
the current state, where for example Erlang/Elixir are no longer
available upstream as documented, so better to have two sets of
instructions for redundancy.

* Improve the separate Grafana server explanation

* Rewrap the text

* Further improve the remot grafana URL

* docs: update changelog

---------

Co-authored-by: Jakob Lichterfeld <jakob-lichterfeld@gmx.de>